+++ admin/lib/accounts/find-all.js
@@ -1,7 +1,7 @@
 module.exports = findAll
 
-var request = require('../../../helpers/request')
-var deserialise = require('../../../helpers/deserialise')
+var request = require('../../../utils/request')
+var deserialise = require('../../../utils/deserialise')
 
 function findAll (state, options) {
   return request({
