+++ admin/lib/accounts/update.js
@@ -1,8 +1,8 @@
 module.exports = update
 
-var request = require('../../../helpers/request')
-var deserialise = require('../../../helpers/deserialise')
-var serialise = require('../../../helpers/serialise')
+var request = require('../../../utils/request')
+var deserialise = require('../../../utils/deserialise')
+var serialise = require('../../../utils/serialise')
 
 function update (state, id, account, options) {
   return request({
