+++ lib/fetch.js
@@ -1,9 +1,9 @@
 module.exports = fetch
 
-var request = require('../helpers/request')
-var saveSession = require('../helpers/save-session')
+var request = require('../utils/request')
+var saveSession = require('../utils/save-session')
 var get = require('./get')
-var deserialise = require('../helpers/deserialise')
+var deserialise = require('../utils/deserialise')
 
 function fetch (state, basePath, path) {
   return request({
