+++ lib/signin.js
@@ -1,10 +1,10 @@
 module.exports = signIn
 
-var request = require('../helpers/request')
-var Promise = require('../helpers/promise')
-var saveSession = require('../helpers/save-session')
-var deserialise = require('../helpers/deserialise')
-var serialise = require('../helpers/serialise')
+var request = require('../utils/request')
+var Promise = require('../utils/promise')
+var saveSession = require('../utils/save-session')
+var deserialise = require('../utils/deserialise')
+var serialise = require('../utils/serialise')
 
 function signIn (state, options) {
   if (!options || !options.username || !options.password) {
