+++ lib/signout.js
@@ -1,7 +1,7 @@
 module.exports = signOut
 
-var request = require('../helpers/request')
-var clearSession = require('../helpers/clear-session')
+var request = require('../utils/request')
+var clearSession = require('../utils/clear-session')
 
 function signOut (state) {
   return request({
