+++ lib/update-profile.js
@@ -1,9 +1,9 @@
 module.exports = updateProfile
 
-var request = require('../helpers/request')
-var Promise = require('../helpers/promise')
-var saveSession = require('../helpers/save-session')
-var serialise = require('../helpers/serialise')
+var request = require('../utils/request')
+var Promise = require('../utils/promise')
+var saveSession = require('../utils/save-session')
+var serialise = require('../utils/serialise')
 
 function updateProfile (state, options) {
   if (!options) {
